[cli] move CLIContext, Command and  ParsedCommand into their own class-files
